move device related classes into model/device
