+++ lib/model/device/device_type.dart
