+++ lib/widgets/pages/device_form.dart
@@ -3,11 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:weforza/exceptions/exceptions.dart';
 import 'package:weforza/generated/l10n.dart';
-import 'package:weforza/model/device.dart';
-import 'package:weforza/model/device_form_delegate.dart';
-import 'package:weforza/model/device_model.dart';
-import 'package:weforza/model/device_type.dart';
-import 'package:weforza/model/device_validator.dart';
+import 'package:weforza/model/device/device.dart';
+import 'package:weforza/model/device/device_form_delegate.dart';
+import 'package:weforza/model/device/device_model.dart';
+import 'package:weforza/model/device/device_type.dart';
+import 'package:weforza/model/device/device_validator.dart';
 import 'package:weforza/riverpod/member/selected_member_devices_provider.dart';
 import 'package:weforza/widgets/common/form_submit_button.dart';
 import 'package:weforza/widgets/common/generic_error.dart';
